Endret navn fra wonderwall til sidecar. Oppdaterte README.
