+++ settings.gradle.kts
@@ -1,6 +1,6 @@
 rootProject.name = "tms-ktor-token-support"
 include("token-support-idporten")
-include("token-support-idporten-wonderwall")
+include("token-support-idporten-sidecar")
 include("token-support-tokendings-exchange")
 include("token-support-tokenx-validation")
 include("token-support-azure-validation")
