+++ token-support-idporten-sidecar/src/main/kotlin/no/nav/tms/token/support/idporten/sidecar/authentication/AuthConfiguration.kt
@@ -1,0 +1,19 @@
+package no.nav.tms.token.support.idporten.sidecar.authentication
+
+import com.auth0.jwk.JwkProvider
+
+internal class AuthConfiguration(
+        jwkProvider: JwkProvider,
+        clientId: String,
+        issuer: String,
+        loginLevel: Int,
+        val fallbackTokenCookieEnabled: Boolean,
+        val fallbackTokenCookieName: String
+) {
+        val tokenVerifier = TokenVerifierBuilder.buildTokenVerifier(
+                jwkProvider = jwkProvider,
+                clientId = clientId,
+                issuer = issuer,
+                loginLevel = loginLevel
+        )
+}
