+++ token-support-idporten-sidecar/src/main/kotlin/no/nav/tms/token/support/idporten/sidecar/authentication/IdPortenTokenPrincipal.kt
@@ -1,0 +1,9 @@
+package no.nav.tms.token.support.idporten.sidecar.authentication
+
+import com.auth0.jwt.interfaces.DecodedJWT
+import io.ktor.auth.*
+
+internal data class IdPortenTokenPrincipal(
+        val accessToken: DecodedJWT
+) : Principal
+
