+++ token-support-idporten-sidecar/src/main/kotlin/no/nav/tms/token/support/idporten/sidecar/authentication/LoginStatus.kt
@@ -1,0 +1,14 @@
+package no.nav.tms.token.support.idporten.sidecar.authentication
+
+import kotlinx.serialization.Serializable
+
+@Serializable
+internal data class LoginStatus(
+    val authenticated: Boolean,
+    val level: Int?
+) {
+    companion object {
+        fun unauthenticated() = LoginStatus(false, null)
+        fun authenticated(level: Int) = LoginStatus(true, level)
+    }
+}
