+++ token-support-idporten-sidecar/src/main/kotlin/no/nav/tms/token/support/idporten/sidecar/authentication/TokenVerifierBuilder.kt
@@ -1,0 +1,18 @@
+package no.nav.tms.token.support.idporten.sidecar.authentication
+
+import com.auth0.jwk.JwkProvider
+
+// Allow easier mocking of TokenVerifier
+internal object TokenVerifierBuilder {
+    fun buildTokenVerifier(
+        jwkProvider: JwkProvider,
+        clientId: String,
+        issuer: String,
+        loginLevel: Int,
+    ) = TokenVerifier(
+        jwkProvider = jwkProvider,
+        clientId = clientId,
+        issuer = issuer,
+        minLoginLevel = loginLevel
+    )
+}
