+++ token-support-idporten-sidecar/src/main/kotlin/no/nav/tms/token/support/idporten/sidecar/authentication/config/constants.kt
@@ -1,0 +1,5 @@
+package no.nav.tms.token.support.idporten.sidecar.authentication.config
+
+internal object Idporten {
+    const val postLoginRedirectCookie = "redirect_uri"
+}
