+++ token-support-idporten-sidecar/src/test/kotlin/no/nav/tms/token/support/idporten/sidecar/JwkBuilder.kt
@@ -1,0 +1,18 @@
+package no.nav.tms.token.support.idporten.sidecar
+
+import com.nimbusds.jose.JWSAlgorithm
+import com.nimbusds.jose.jwk.KeyUse
+import com.nimbusds.jose.jwk.RSAKey
+import com.nimbusds.jose.jwk.gen.RSAKeyGenerator
+
+object JwkBuilder {
+    fun generateJwk(): RSAKey {
+        return RSAKeyGenerator(2048)
+                .keyUse(KeyUse.SIGNATURE)
+                .algorithm(JWSAlgorithm.RS256)
+                .keyID("KID")
+                .generate()
+    }
+
+    fun generateJwkString(): String = generateJwk().toJSONString()
+}
