+++ token-support-idporten-sidecar/src/test/kotlin/no/nav/tms/token/support/idporten/sidecar/ObjectMapper.kt
@@ -1,0 +1,9 @@
+package no.nav.tms.token.support.idporten.sidecar
+
+import kotlinx.serialization.json.Json
+
+internal object ObjectMapper {
+    val objectMapper = Json {
+        ignoreUnknownKeys = true
+    }
+}
