--- token-support-idporten-wonderwall/src/main/kotlin/no/nav/tms/token/support/idporten/wonderwall/authentication/IdPortenTokenPrincipal.kt
@@ -1,9 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall.authentication
-
-import com.auth0.jwt.interfaces.DecodedJWT
-import io.ktor.auth.*
-
-internal data class IdPortenTokenPrincipal(
-        val accessToken: DecodedJWT
-) : Principal
-
