--- token-support-idporten-wonderwall/src/main/kotlin/no/nav/tms/token/support/idporten/wonderwall/authentication/LoginStatus.kt
@@ -1,14 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall.authentication
-
-import kotlinx.serialization.Serializable
-
-@Serializable
-internal data class LoginStatus(
-    val authenticated: Boolean,
-    val level: Int?
-) {
-    companion object {
-        fun unauthenticated() = LoginStatus(false, null)
-        fun authenticated(level: Int) = LoginStatus(true, level)
-    }
-}
