--- token-support-idporten-wonderwall/src/main/kotlin/no/nav/tms/token/support/idporten/wonderwall/authentication/OauthServerConfigurationMetadata.kt
@@ -1,12 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall.authentication
-
-import kotlinx.serialization.SerialName
-import kotlinx.serialization.Serializable
-
-// This object holds info returned from idporten's well-known-url
-@Serializable
-internal data class OauthServerConfigurationMetadata(
-        @SerialName("issuer") val issuer: String,
-        @SerialName("token_endpoint") val tokenEndpoint: String,
-        @SerialName("jwks_uri") val jwksUri: String,
-)
