--- token-support-idporten-wonderwall/src/main/kotlin/no/nav/tms/token/support/idporten/wonderwall/authentication/config/constants.kt
@@ -1,5 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall.authentication.config
-
-internal object Idporten {
-    const val postLoginRedirectCookie = "redirect_uri"
-}
