--- token-support-idporten-wonderwall/src/main/kotlin/no/nav/tms/token/support/idporten/wonderwall/authentication/config/httpClient.kt
@@ -1,18 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall.authentication.config
-
-import io.ktor.client.*
-import io.ktor.client.request.*
-import io.ktor.http.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import no.nav.tms.token.support.idporten.wonderwall.authentication.OauthServerConfigurationMetadata
-
-
-internal suspend fun HttpClient.getOAuthServerConfigurationMetadata(url: String)
-        : OauthServerConfigurationMetadata = withContext(Dispatchers.IO) {
-    request {
-        method = HttpMethod.Get
-        url(url)
-        accept(ContentType.Application.Json)
-    }
-}
