--- token-support-idporten-wonderwall/src/test/kotlin/no/nav/tms/token/support/idporten/wonderwall/JwkBuilder.kt
@@ -1,18 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall
-
-import com.nimbusds.jose.JWSAlgorithm
-import com.nimbusds.jose.jwk.KeyUse
-import com.nimbusds.jose.jwk.RSAKey
-import com.nimbusds.jose.jwk.gen.RSAKeyGenerator
-
-object JwkBuilder {
-    fun generateJwk(): RSAKey {
-        return RSAKeyGenerator(2048)
-                .keyUse(KeyUse.SIGNATURE)
-                .algorithm(JWSAlgorithm.RS256)
-                .keyID("KID")
-                .generate()
-    }
-
-    fun generateJwkString(): String = generateJwk().toJSONString()
-}
