--- token-support-idporten-wonderwall/src/test/kotlin/no/nav/tms/token/support/idporten/wonderwall/ObjectMapper.kt
@@ -1,9 +1,0 @@
-package no.nav.tms.token.support.idporten.wonderwall
-
-import kotlinx.serialization.json.Json
-
-internal object ObjectMapper {
-    val objectMapper = Json {
-        ignoreUnknownKeys = true
-    }
-}
